Move dataarrays creation to parsed2zarr
