+++ echopype/convert/set_groups_base.py
@@ -1,7 +1,6 @@
 import abc
-from typing import List, Set, Tuple
+from typing import Set
 
-import dask.array
 import numpy as np
 import pynmea2
 import xarray as xr
@@ -51,7 +50,7 @@ class SetGroupsBase(abc.ABC):
         else:
             self.compression_settings = COMPRESSION_SETTINGS[self.engine]
 
-        self._varattrs = sonarnetcdf_1.yaml_dict["variable_and_varattributes"]
+        self._varattrs = self.parsed2zarr_obj._varattrs
         # self._beamgroups must be a list of dicts, eg:
         # [{"name":"Beam_group1", "descr":"contains complex backscatter data
         # and other beam or channel-specific data."}]
@@ -331,229 +330,3 @@ class SetGroupsBase(abc.ABC):
 
         self._add_ping_time_dim(ds, beam_ping_time_names, ping_time_only_names)
         self._add_beam_dim(ds, beam_only_names, beam_ping_time_names)
-
-    def _get_channel_ids(self, chan_str: np.ndarray) -> List[str]:
-        """
-        Obtains the channel IDs associated with ``chan_str``.
-
-        Parameters
-        ----------
-        chan_str : np.ndarray
-            A numpy array of strings corresponding to the
-            keys of ``config_datagram["transceivers"]``
-
-        Returns
-        -------
-        A list of strings representing the channel IDS
-        """
-        if self.sonar_model in ["EK60", "ES70"]:
-            return [
-                self.parser_obj.config_datagram["transceivers"][int(i)]["channel_id"]
-                for i in chan_str
-            ]
-        else:
-            return [
-                self.parser_obj.config_datagram["configuration"][i]["channel_id"] for i in chan_str
-            ]
-
-    def _get_power_dataarray(self, zarr_path: str) -> xr.DataArray:
-        """
-        Constructs a DataArray from a Dask array for the power
-        data.
-
-        Parameters
-        ----------
-        zarr_path: str
-            Path to the zarr file that contain the power data
-
-        Returns
-        -------
-        DataArray named "backscatter_r" representing the
-        power data.
-        """
-
-        # collect variables associated with the power data
-        power = dask.array.from_zarr(zarr_path, component="power/power")
-
-        pow_time_path = "power/" + self.parsed2zarr_obj.power_dims[0]
-        pow_chan_path = "power/" + self.parsed2zarr_obj.power_dims[1]
-        power_time = dask.array.from_zarr(zarr_path, component=pow_time_path).compute()
-        power_channel = dask.array.from_zarr(zarr_path, component=pow_chan_path).compute()
-
-        # obtain channel names for power data
-        pow_chan_names = self._get_channel_ids(power_channel)
-
-        backscatter_r = xr.DataArray(
-            data=power,
-            coords={
-                "ping_time": (
-                    ["ping_time"],
-                    power_time,
-                    self._varattrs["beam_coord_default"]["ping_time"],
-                ),
-                "channel": (
-                    ["channel"],
-                    pow_chan_names,
-                    self._varattrs["beam_coord_default"]["channel"],
-                ),
-                "range_sample": (
-                    ["range_sample"],
-                    np.arange(power.shape[2]),
-                    self._varattrs["beam_coord_default"]["range_sample"],
-                ),
-            },
-            name="backscatter_r",
-            attrs={
-                "long_name": self._varattrs["beam_var_default"]["backscatter_r"]["long_name"],
-                "units": "dB",
-            },
-        )
-
-        return backscatter_r
-
-    def _get_angle_dataarrays(self, zarr_path: str) -> Tuple[xr.DataArray, xr.DataArray]:
-        """
-        Constructs the DataArrays from Dask arrays associated
-        with the angle data.
-
-        Parameters
-        ----------
-        zarr_path: str
-            Path to the zarr file that contains the angle data
-
-        Returns
-        -------
-        DataArrays named "angle_athwartship" and "angle_alongship",
-        respectively, representing the angle data.
-        """
-
-        # collect variables associated with the angle data
-        angle_along = dask.array.from_zarr(zarr_path, component="angle/angle_alongship")
-        angle_athwart = dask.array.from_zarr(zarr_path, component="angle/angle_athwartship")
-
-        ang_time_path = "angle/" + self.parsed2zarr_obj.angle_dims[0]
-        ang_chan_path = "angle/" + self.parsed2zarr_obj.angle_dims[1]
-        angle_time = dask.array.from_zarr(zarr_path, component=ang_time_path).compute()
-        angle_channel = dask.array.from_zarr(zarr_path, component=ang_chan_path).compute()
-
-        # obtain channel names for angle data
-        ang_chan_names = self._get_channel_ids(angle_channel)
-
-        array_coords = {
-            "ping_time": (
-                ["ping_time"],
-                angle_time,
-                self._varattrs["beam_coord_default"]["ping_time"],
-            ),
-            "channel": (
-                ["channel"],
-                ang_chan_names,
-                self._varattrs["beam_coord_default"]["channel"],
-            ),
-            "range_sample": (
-                ["range_sample"],
-                np.arange(angle_athwart.shape[2]),
-                self._varattrs["beam_coord_default"]["range_sample"],
-            ),
-        }
-
-        angle_athwartship = xr.DataArray(
-            data=angle_athwart,
-            coords=array_coords,
-            name="angle_athwartship",
-            attrs={
-                "long_name": "electrical athwartship angle",
-                "comment": (
-                    "Introduced in echopype for Simrad echosounders. "  # noqa
-                    + "The athwartship angle corresponds to the major angle in SONAR-netCDF4 vers 2. "  # noqa
-                ),
-            },
-        )
-
-        angle_alongship = xr.DataArray(
-            data=angle_along,
-            coords=array_coords,
-            name="angle_alongship",
-            attrs={
-                "long_name": "electrical alongship angle",
-                "comment": (
-                    "Introduced in echopype for Simrad echosounders. "  # noqa
-                    + "The alongship angle corresponds to the minor angle in SONAR-netCDF4 vers 2. "  # noqa
-                ),
-            },
-        )
-
-        return angle_athwartship, angle_alongship
-
-    def _get_complex_dataarrays(self, zarr_path: str) -> Tuple[xr.DataArray, xr.DataArray]:
-        """
-        Constructs the DataArrays from Dask arrays associated
-        with the complex data.
-
-        Parameters
-        ----------
-        zarr_path: str
-            Path to the zarr file that contains the complex data
-
-        Returns
-        -------
-        DataArrays named "backscatter_r" and "backscatter_i",
-        respectively, representing the complex data.
-        """
-
-        # collect variables associated with the complex data
-        complex_r = dask.array.from_zarr(zarr_path, component="complex/backscatter_r")
-        complex_i = dask.array.from_zarr(zarr_path, component="complex/backscatter_i")
-
-        comp_time_path = "complex/" + self.parsed2zarr_obj.complex_dims[0]
-        comp_chan_path = "complex/" + self.parsed2zarr_obj.complex_dims[1]
-        complex_time = dask.array.from_zarr(zarr_path, component=comp_time_path).compute()
-        complex_channel = dask.array.from_zarr(zarr_path, component=comp_chan_path).compute()
-
-        # obtain channel names for complex data
-        comp_chan_names = self._get_channel_ids(complex_channel)
-
-        array_coords = {
-            "ping_time": (
-                ["ping_time"],
-                complex_time,
-                self._varattrs["beam_coord_default"]["ping_time"],
-            ),
-            "channel": (
-                ["channel"],
-                comp_chan_names,
-                self._varattrs["beam_coord_default"]["channel"],
-            ),
-            "range_sample": (
-                ["range_sample"],
-                np.arange(complex_r.shape[2]),
-                self._varattrs["beam_coord_default"]["range_sample"],
-            ),
-            "beam": (
-                ["beam"],
-                np.arange(start=1, stop=complex_r.shape[3] + 1).astype(str),
-                self._varattrs["beam_coord_default"]["beam"],
-            ),
-        }
-
-        backscatter_r = xr.DataArray(
-            data=complex_r,
-            coords=array_coords,
-            name="backscatter_r",
-            attrs={
-                "long_name": self._varattrs["beam_var_default"]["backscatter_r"]["long_name"],
-                "units": "dB",
-            },
-        )
-
-        backscatter_i = xr.DataArray(
-            data=complex_i,
-            coords=array_coords,
-            name="backscatter_i",
-            attrs={
-                "long_name": self._varattrs["beam_var_default"]["backscatter_i"]["long_name"],
-                "units": "dB",
-            },
-        )
-
-        return backscatter_r, backscatter_i
